Move markdown_parser into lib, update views and templates
